Add sponsor card for Fabiano Abucarub - Perito
Added a new sponsor card to the HTML file for "Fabiano Abucarub - Perito."
This includes an `<a>` tag with attributes such as `href`, `class`, `data-full`,
`data-link`, and `title`, as well as an `<img>` tag to display the sponsor's
image (`6oBurger/FabianoPerito.png`). The card links to the sponsor's Instagram
profile and ensures proper display on the webpage.