Replace sponsor image and update index.html
The binary file `casadasfraldaslondrina.png` was removed and replaced with a new image file, `casadasfraldaslondrina.jpg`.

In `index.html`, a new sponsor card was added for "Tk Acabamentos Londrina," including an `<a>` tag with relevant attributes and an `<img>` tag pointing to the new image. Minor spacing adjustments were made to accommodate the changes.