Add sponsor cards for Automaq and Medclin
Added a sponsor card for "Automaq Automação Comercial" with a link to their Instagram page and an image.
Added a sponsor card for "Medclin Ocupacional" with a link to their Facebook page and an image.
These changes enhance the webpage by including new sponsors and their respective links and images.