Add new sponsor cards and fix formatting issue
Added a new sponsor card for "Morcelli Cervejaria" with an image
and a link to their Instagram page. Added another sponsor card
for "SinPro Londrina" with an image and a link to their Instagram
page. Fixed a formatting issue with the closing `</a>` tag for
the "SinPro Londrina" sponsor card to ensure proper structure.